feat: change return type of SetOrganizationsRolesIdForUserIdAsync method from long to int

+++ RFRGOBAC/IServices/IRoleXUserXOrganizationService.cs
@@ -18,7 +18,7 @@ public interface IRoleXUserXOrganizationService : ICommonJoinService<RoleXUserXO
     Task<IEnumerable<long>> GetRolesIdByUsersIdAndOrganizationIdAsync(IEnumerable<long> usersId, long OrganizationId, RoleXUserXOrganizationQueryOptions? options = null);
     Task<IEnumerable<long>> GetAllRolesIdByUsersIdAndOrganizationIdAsync(IEnumerable<long> usersId, long OrganizationId, RoleXUserXOrganizationQueryOptions? options = null);
     Task<IEnumerable<Organization>> GetOrganizationsByUsersIdAsync(IEnumerable<long> usersId, RoleXUserXOrganizationQueryOptions? options = null);
-    Task<long> SetOrganizationsRolesIdForUserIdAsync(IEnumerable<OrganizationRolesId> organizationsRolesId, long userId, RoleXUserXOrganizationQueryOptions? options = null);
+    Task<int> SetOrganizationsRolesIdForUserIdAsync(IEnumerable<OrganizationRolesId> organizationsRolesId, long userId, RoleXUserXOrganizationQueryOptions? options = null);
     Task<IEnumerable<OrganizationRoles>> GetOrganizationsRolesByUserIdAsync(long userId, RoleXUserXOrganizationQueryOptions? options = null);
     Task<IEnumerable<Organization>> GetOrganizationsByUserIdAsync(long userId, RoleXUserXOrganizationQueryOptions? options = null);
 }
